refactor: Update product prices modal actions to use new context and type

+++ client/src/features/product/__test__/components/modal/DeleteProdPrice.tsx
@@ -1,40 +1,45 @@
 import { UseModalProps } from '@/utils/Modal';
-import { useProductPrices } from '../..';
 import { useProductPricesMutation } from '../../hooks';
 import { useState } from 'react';
 import { Button } from '@/components';
 import { toast } from 'react-toastify';
+import { useProductPricesPaginated } from '../../context/ProductPricesPaginatedContext';
 
 interface DeleteProdPriceProps {
 	onClose: UseModalProps['closeModal'];
 }
 
 export const DeleteProdPrice = ({ onClose }: DeleteProdPriceProps) => {
-	const { selectedProductPrice } = useProductPrices();
+	const { selectedProductPrice } = useProductPricesPaginated();
+
 	const { deleteProductPricesMutation } = useProductPricesMutation();
 
 	const [isSubmitting, setIsSubmitting] = useState<boolean>(false);
 	const handleDelete = () => {
-		setIsSubmitting(true);
-		deleteProductPricesMutation(selectedProductPrice.id)
-			.then(res => {
-				toast.success(res.message);
-				onClose();
-			})
-			.catch(error => {
-				toast.error(
-					error.message
-						? error.message
-						: 'Failed to delete product listing.',
-				);
-			})
-			.finally(() => {
-				setIsSubmitting(false);
-			});
+		if (selectedProductPrice) {
+			setIsSubmitting(true);
+			deleteProductPricesMutation(selectedProductPrice.id)
+				.then(res => {
+					toast.success(res.message);
+					onClose();
+				})
+				.catch(error => {
+					toast.error(
+						error.message
+							? error.message
+							: 'Failed to delete product listing.',
+					);
+				})
+				.finally(() => {
+					setIsSubmitting(false);
+				});
+		} else {
+			toast.error('No product selected.');
+		}
 	};
 
 	return (
-		<div className="flex max-w-md flex-col gap-4">
+		<div className="flex max-w-lg flex-col gap-4">
 			<h2 className="text-sm font-medium text-slate-800">
 				Are you sure you would like to delete this item? This action{' '}
 				<strong>cannot</strong>&nbsp; be undone.
@@ -43,35 +48,53 @@ export const DeleteProdPrice = ({ onClose }: DeleteProdPriceProps) => {
 			<div className="mt-3 grid w-full grid-flow-row grid-cols-12 gap-3">
 				<div className="col-span-3 flex flex-col justify-center gap-1">
 					<h3 className="text-sm font-bold text-gray-600">Name</h3>
-					<p className="text-sm">{selectedProductPrice.product.name}</p>
+					<p className="text-sm">
+						{selectedProductPrice?.product.name ?? 'No name'}
+					</p>
 				</div>
 				<div className="col-span-3 flex flex-col justify-center gap-1">
 					<h3 className="text-sm font-bold text-gray-600">Brand</h3>
 					<p className="text-sm">
-						{selectedProductPrice.product.brand || 'No brand'}
+						{selectedProductPrice?.product.brand || 'No brand'}
 					</p>
 				</div>
-				<div className="col-span-6 flex flex-col justify-center gap-1">
+				<div className="col-span-3 flex flex-col justify-center gap-1">
 					<h3 className="text-sm font-bold text-gray-600">Serial No.</h3>
 					<p className="text-sm">
-						{selectedProductPrice.product.serial_no}
+						{selectedProductPrice?.product.serial_no ??
+							'No serial number'}
+					</p>
+				</div>
+				<div className="col-span-3 flex flex-col justify-center	gap-1">
+					<h3 className="text-sm font-bold text-gray-600">Inventory</h3>
+					<p className="text-sm">
+						{selectedProductPrice?.inventory_product.inventory.code ??
+							'No inventory'}
 					</p>
 				</div>
 				<div className="col-span-3 flex flex-col justify-center gap-1">
 					<h3 className="text-sm font-bold text-gray-600">Size</h3>
-					<p className="text-sm">{selectedProductPrice.product.size}</p>
+					<p className="text-sm">
+						{selectedProductPrice?.product.size ?? 'No size'}
+					</p>
 				</div>
 				<div className="col-span-3 flex flex-col justify-center gap-1">
 					<h3 className="text-sm font-bold text-gray-600">Color</h3>
-					<p className="text-sm">{selectedProductPrice.product.color}</p>
+					<p className="text-sm">
+						{selectedProductPrice?.product.color ?? 'No color'}
+					</p>
 				</div>
 				<div className="col-span-3 flex flex-col justify-center	gap-1">
 					<h3 className="text-sm font-bold text-gray-600">Unit</h3>
-					<p className="text-sm">{selectedProductPrice.unit}</p>
+					<p className="text-sm">
+						{selectedProductPrice?.unit ?? 'No unit'}
+					</p>
 				</div>
 				<div className="col-span-3 flex flex-col justify-center	gap-1">
 					<h3 className="text-sm font-bold text-gray-600">Warehouse</h3>
-					<p className="text-sm">{selectedProductPrice.warehouse.code}</p>
+					<p className="text-sm">
+						{selectedProductPrice?.warehouse.code ?? 'No warehouse'}
+					</p>
 				</div>
 				<div className="col-span-4 flex flex-col justify-center	gap-1">
 					<h3 className="text-sm font-bold text-gray-600">
@@ -81,7 +104,7 @@ export const DeleteProdPrice = ({ onClose }: DeleteProdPriceProps) => {
 						{new Intl.NumberFormat('en-US', {
 							style: 'currency',
 							currency: 'PHP',
-						}).format(selectedProductPrice.price)}
+						}).format(selectedProductPrice?.price ?? 0)}
 					</p>
 				</div>
 				<div className="col-span-4 flex flex-col justify-center	gap-1">
@@ -90,7 +113,7 @@ export const DeleteProdPrice = ({ onClose }: DeleteProdPriceProps) => {
 						{new Intl.NumberFormat('en-US', {
 							style: 'currency',
 							currency: 'PHP',
-						}).format(selectedProductPrice.markup_price)}
+						}).format(selectedProductPrice?.markup_price ?? 0)}
 					</p>
 				</div>
 				<div className="col-span-4 flex flex-col justify-center	gap-1">
@@ -99,7 +122,7 @@ export const DeleteProdPrice = ({ onClose }: DeleteProdPriceProps) => {
 						{new Intl.NumberFormat('en-US', {
 							style: 'currency',
 							currency: 'PHP',
-						}).format(selectedProductPrice.price)}
+						}).format(selectedProductPrice?.price ?? 0)}
 					</p>
 				</div>
 			</div>
